Example of how to use Pulp 3 plugin API
The Example plugin for Pulp 3 manages ExampleContent. ExampleContent repositories are the
same as the ISO repositories in Pulp 2. Any type of file can be uploaded to an Example
repository.

The Example plugin provides two importers: ExampleFuturesImporter and ExampleAsyncIOImporter.
Both importers require the feed URL to point to a PULP_MANIFEST inside of a remote
repository. The importers demonstrate how to download content from a remote repository, create
content in Pulp, and associate that content with a repository. The ExampleFuturesImporter uses
pulpcore.plugin.download.futures.Batch to download files concurrently. The ExampleAsyncIOImporter
uses pulpcore.plugin.download.asyncio.GroupDownloader to perform downloads. The rest of the
implementation is the same for both importers.

The Example plugin also provides an ExamplePublisher. The ExamplePublisher demonstrates
how the plugin API can be used to create Publications from content associated with a repository.